use custom neural net
