+++ zeromq_server/src/main/kotlin/StartZeroMQServer.kt
@@ -2,7 +2,7 @@ package org.example
 
 import io.vertx.core.Vertx
 import io.vertx.core.VertxOptions
-import org.deep_thinker.agent.dqn.djl.DQNAgentFactoryVerticle
+import org.deep_thinker.agent.dqn.deep_thinker_dl.DQNAgentFactoryVerticle
 import org.example.org.deep_thinker.zeromq.server.DeepThinkerZeroMQServer
 
 fun main() {
